refactor: moved ws acceptance test to ws-server package
