+++ packages/ws-server/tests/acceptance/blockNumber.spec.ts
@@ -21,21 +21,14 @@
 // external resources
 import { expect } from 'chai';
 import { ethers, WebSocketProvider } from 'ethers';
-import RelayClient from '../../clients/relayClient';
+import RelayClient from '@hashgraph/json-rpc-server/tests/clients/relayClient';
 
-describe('@release @web-socket eth_getBlockByHash', async function () {
+describe('@release @web-socket eth_blockNumber', async function () {
   const WS_RELAY_URL = `${process.env.WS_RELAY_URL}`;
-  const METHOD_NAME = 'eth_getBlockByHash';
-  const FAKE_TX_HASH = `0x${'00'.repeat(32)}`;
+  const METHOD_NAME = 'eth_blockNumber';
   const INVALID_PARAMS = [
-    [],
-    [FAKE_TX_HASH],
-    [FAKE_TX_HASH, '0xhbar'],
-    [FAKE_TX_HASH, '54'],
-    [FAKE_TX_HASH, true, 39],
-    [FAKE_TX_HASH, false, 39],
-    ['0xhedera', true],
-    ['0xhbar', false],
+    ['hedera', 'hbar'],
+    ['websocket', 'rpc', 'invalid'],
   ];
 
   let relayClient: RelayClient, wsProvider: WebSocketProvider;
@@ -43,7 +36,6 @@ describe('@release @web-socket eth_getBlockByHash', async function () {
   before(async () => {
     // @ts-ignore
     const { relay } = global;
-
     relayClient = relay;
   });
 
@@ -70,8 +62,9 @@ describe('@release @web-socket eth_getBlockByHash', async function () {
   }
 
   it('Should handle valid requests correctly', async () => {
-    const expectedResult = await relayClient.call('eth_getBlockByNumber', ['latest', true]);
-    const result = await wsProvider.send(METHOD_NAME, [expectedResult.hash, true]);
-    expect(result).to.deep.eq(expectedResult);
+    const result = await wsProvider.send(METHOD_NAME, []);
+    const expectedResult = await relayClient.call(METHOD_NAME, []);
+
+    expect(result).to.eq(expectedResult);
   });
 });
