--- test-results.02088f1d2a29bfef82062b5ed525526f.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<testsuites name="Mocha Tests" time="3.203" tests="7" failures="0">
-  <testsuite name="Root Suite" timestamp="2024-04-11T16:01:50" tests="0" time="0.000" failures="0">
-  </testsuite>
-  <testsuite name="RPC Server Acceptance Tests" timestamp="2024-04-11T16:01:50" tests="0" file="/Users/user/development/hashgraph/hedera-json-rpc-relay/packages/server/tests/acceptance/index.spec.ts" time="0.000" failures="0">
-  </testsuite>
-  <testsuite name="Acceptance tests" timestamp="2024-04-11T16:01:50" tests="0" file="/Users/user/development/hashgraph/hedera-json-rpc-relay/packages/server/tests/acceptance/index.spec.ts" time="0.000" failures="0">
-  </testsuite>
-  <testsuite name="@release @web-socket eth_getLogs" timestamp="2024-04-11T16:01:50" tests="7" file="/Users/user/development/hashgraph/hedera-json-rpc-relay/packages/server/tests/acceptance/index.spec.ts" time="3.175" failures="0">
-    <testcase name="RPC Server Acceptance Tests Acceptance tests @release @web-socket eth_getLogs Should throw predefined.INVALID_PARAMETERS if the request&apos;s params variable is invalid (params.length !== 1). params=[[object Object],]" time="0.011" classname="Should throw predefined.INVALID_PARAMETERS if the request&apos;s params variable is invalid (params.length !== 1). params=[[object Object],]">
-    </testcase>
-    <testcase name="RPC Server Acceptance Tests Acceptance tests @release @web-socket eth_getLogs Should throw predefined.INVALID_PARAMETERS if the request&apos;s params variable is invalid (params.length !== 1). params=[]" time="0.003" classname="Should throw predefined.INVALID_PARAMETERS if the request&apos;s params variable is invalid (params.length !== 1). params=[]">
-    </testcase>
-    <testcase name="RPC Server Acceptance Tests Acceptance tests @release @web-socket eth_getLogs Should throw predefined.INVALID_PARAMETERS if the request&apos;s params variable is invalid (params.length !== 1). params=[[object Object],0xhbar,false]" time="0.001" classname="Should throw predefined.INVALID_PARAMETERS if the request&apos;s params variable is invalid (params.length !== 1). params=[[object Object],0xhbar,false]">
-    </testcase>
-    <testcase name="RPC Server Acceptance Tests Acceptance tests @release @web-socket eth_getLogs Should handle invalid data correctly. filter = {&quot;address&quot;:&quot;0xhedera&quot;,&quot;fromBlock&quot;:&quot;latest&quot;,&quot;toBlock&quot;:&quot;latest&quot;}" time="0.042" classname="Should handle invalid data correctly. filter = {&quot;address&quot;:&quot;0xhedera&quot;,&quot;fromBlock&quot;:&quot;latest&quot;,&quot;toBlock&quot;:&quot;latest&quot;}">
-    </testcase>
-    <testcase name="RPC Server Acceptance Tests Acceptance tests @release @web-socket eth_getLogs Should handle invalid data correctly. filter = {&quot;address&quot;:&quot;0x637a6A8e5A69C087c24983B05261F63f64ED7e9b&quot;,&quot;fromBlock&quot;:&quot;0xhedera&quot;,&quot;toBlock&quot;:&quot;latest&quot;}" time="0.005" classname="Should handle invalid data correctly. filter = {&quot;address&quot;:&quot;0x637a6A8e5A69C087c24983B05261F63f64ED7e9b&quot;,&quot;fromBlock&quot;:&quot;0xhedera&quot;,&quot;toBlock&quot;:&quot;latest&quot;}">
-    </testcase>
-    <testcase name="RPC Server Acceptance Tests Acceptance tests @release @web-socket eth_getLogs Should handle invalid data correctly. filter = {&quot;address&quot;:&quot;0x637a6A8e5A69C087c24983B05261F63f64ED7e9b&quot;,&quot;fromBlock&quot;:&quot;latest&quot;,&quot;toBlock&quot;:&quot;0xhedera&quot;}" time="0.005" classname="Should handle invalid data correctly. filter = {&quot;address&quot;:&quot;0x637a6A8e5A69C087c24983B05261F63f64ED7e9b&quot;,&quot;fromBlock&quot;:&quot;latest&quot;,&quot;toBlock&quot;:&quot;0xhedera&quot;}">
-    </testcase>
-    <testcase name="RPC Server Acceptance Tests Acceptance tests @release @web-socket eth_getLogs Should handle valid data correctly" time="1.515" classname="Should handle valid data correctly">
-    </testcase>
-  </testsuite>
-</testsuites>
